Reorganise files, prepare for command migration
