+++ commandsOLD/8ball.js
