+++ commandsOLD/bite.js
