+++ commandsOLD/coinflip.js
