+++ commandsOLD/diceroll.js
