+++ commandsOLD/kill.js
