+++ commandsOLD/lick.js
