+++ commandsOLD/smug.js
