+++ events/ready.js
@@ -1,7 +1,9 @@
+const { Events } = require('discord.js');
+
 module.exports = {
 	name: 'ready',
 	once: true,
 	execute(client) {
-		console.log(`Ready! Riru is now online! Logged in as ${client.user.tag}`);
+		console.log(`Riru is now online! Logged in as ${client.user.tag}`);
 	},
 };
